+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 01. :gift: Added exception `dev.orne.http.client.HttpClientException`
 01. :gift: Added exception `dev.orne.http.client.HttpResponseException`
 01. :gift: Added interface `dev.orne.http.client.StatusIndependientOperation`
+01. :gift: Added class `dev.orne.http.client.BaseHttpServiceClient`